requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.5
py==1.7.0
pylint==2.1.1
pytest==3.9.2
pytest-cov==2.6.0
pytest-django==3.4.3
pytz==2018.6
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.24
wrapt==1.10.11
